[#352]refactor : 펫 정보 리팩토링. controller층 분리, 타입 별도 파일로 분리
